--- .github/workflows/run-access-management-consent-breakpoint-tests.yml
@@ -1,29 +1,0 @@
-name: "access-management - run k6 consent breakpoint tests"
-on:
-  schedule:
-    - cron: "0 08 * * 1"
-run-name: "run breakpoint"
-permissions:
-  contents: "read"
-  id-token: "write"
-jobs:
-  k6-breakpoint:
-    environment: "test"
-    runs-on: "ubuntu-latest"
-    steps:
-      - name: "Checkout code"
-        uses: "actions/checkout@3d3c42e5aac5ba805825da76410c181273ba90b1"
-      - name: "Setup k8s client"
-        uses: "./.github/actions/setup-k8s"
-        with:
-          azure-client-id: "${{ secrets.AZURE_CLIENT_ID }}"
-          azure-subscription-id: "${{ secrets.AZURE_SUBSCRIPTION_ID }}"
-          azure-tenant-id: "${{ secrets.AZURE_TENANT_ID }}"
-      - name: "Generate k8s manifests"
-        uses: "Altinn/altinn-platform/actions/generate-k6-manifests@ec383f5b2cafb06b2d760d14ae2d06f711d8e3e7"
-        with:
-          config_file: "K6/api/tests/access-management/consent/breakpoint.yaml"
-      - if: "github.event.schedule == '0 08 * * 1'"
-        name: "Run Breakpoint for .dist/yt01-post-consent-break"
-        run: "kubectl apply --server-side -f .dist/yt01-post-consent-break"
-        shell: "bash"
