--- K6/api/tests/access-management-bff/breakpoint.yaml
@@ -1,68 +1,0 @@
-namespace: access-management-bff
-base_dir: K6/api/tests/access-management-bff
-test_definitions:
-  - test_scope: access-management-bff
-    test_file: client-admin.js
-    contexts:
-      - environment: yt01
-        test_run:
-          env:
-            - name: BREAKPOINT_STAGE_DURATION
-              value: "10m"
-            - name: BREAKPOINT_STAGE_TARGET
-              value: 100
-
-  - test_scope: access-management-bff
-    test_file: fullmakt-user-user.js
-    contexts:
-      - environment: yt01
-        test_run:
-          env:
-            - name: BREAKPOINT_STAGE_DURATION
-              value: "10m"
-            - name: BREAKPOINT_STAGE_TARGET
-              value: 100
-
-  - test_scope: access-management-bff
-    test_file: fullmakt-org-org.js
-    contexts:
-      - environment: yt01
-        test_run:
-          env:
-            - name: BREAKPOINT_STAGE_DURATION
-              value: "10m"
-            - name: BREAKPOINT_STAGE_TARGET
-              value: 100
-
-  - test_scope: access-management-bff
-    test_file: single-service-org-org.js
-    contexts:
-      - environment: yt01
-        test_run:
-          env:
-            - name: BREAKPOINT_STAGE_DURATION
-              value: "10m"
-            - name: BREAKPOINT_STAGE_TARGET
-              value: 100
-
-  - test_scope: access-management-bff
-    test_file: single-service-user-user.js
-    contexts:
-      - environment: yt01
-        test_run:
-          env:
-            - name: BREAKPOINT_STAGE_DURATION
-              value: "10m"
-            - name: BREAKPOINT_STAGE_TARGET
-              value: 100
-
-  - test_scope: access-management-bff
-    test_file: delegation-export.js
-    contexts:
-      - environment: yt01
-        test_run:
-          env:
-            - name: BREAKPOINT_STAGE_DURATION
-              value: "10m"
-            - name: BREAKPOINT_STAGE_TARGET
-              value: 50
