--- K6/api/tests/access-management-bff/functional.yaml
@@ -1,48 +1,0 @@
-namespace: access-management-bff
-base_dir: K6/api/tests/access-management-bff
-test_definitions:
-  - test_file: client-admin.js
-    contexts:
-      - environment: yt01
-      - environment: at23
-      - environment: tt02
-
-  - test_file: fullmakt-user-user.js
-    contexts:
-      - environment: yt01
-      - environment: at23
-      - environment: tt02
-
-  - test_file: fullmakt-org-org.js
-    contexts:
-      - environment: yt01
-      - environment: at23
-      - environment: tt02
-
-  - test_file: single-service-org-org.js
-    contexts:
-      - environment: yt01
-      - environment: at23
-      - environment: tt02
-
-  - test_file: single-service-user-user.js
-    contexts:
-      - environment: yt01
-      - environment: at23
-      - environment: tt02
-
-  - test_file: instance-delegation-user-user.js
-    contexts:
-      - environment: at23
-      - environment: tt02
-
-  - test_file: instance-delegation-org-user.js
-    contexts:
-      - environment: at23
-      - environment: tt02
-
-  - test_file: delegation-export.js
-    contexts:
-      - environment: yt01
-      - environment: at23
-      - environment: tt02
