+++ K6/api/tests/access-management-bff/instance-delegation/functional.yaml
@@ -1,0 +1,12 @@
+namespace: access-management-bff
+base_dir: K6/api/tests/access-management-bff/instance-delegation
+test_definitions:
+  - test_file: user-to-user.js
+    contexts:
+      - environment: at23
+      - environment: tt02
+
+  - test_file: org-to-user.js
+    contexts:
+      - environment: at23
+      - environment: tt02
