+++ K6/api/tests/access-management-bff/single-rights/smoke.yaml
@@ -1,0 +1,20 @@
+namespace: access-management-bff
+base_dir: K6/api/tests/access-management-bff/single-rights
+test_definitions:
+  - test_file: org-to-org.js
+    contexts:
+      - environment: yt01
+        test_type:
+          type: smoke
+          enabled: true
+          config_file: ../conf.json
+      - environment: at23
+
+  - test_file: user-to-user.js
+    contexts:
+      - environment: yt01
+        test_type:
+          type: smoke
+          enabled: true
+          config_file: ../conf.json
+      - environment: at23
