+++ K6/api/tests/access-management-bff/single-rights/user-to-user.js
@@ -6,22 +6,21 @@ import {
     AccessPackageClient,
     GetAccessPackageDelegationsQueryBuilder,
     SearchAccessPackagesQueryBuilder,
-} from "../../../clients/access-management-bff/access-package/index.js";
-import { AltinnCdnClient } from "../../../clients/access-management-bff/altinn-cdn/index.js";
+} from "../../../../clients/access-management-bff/access-package/index.js";
 import {
     ConnectionClient,
-    CreateRightHolderQueryBuilder,
     DeleteReporteeConnectionQueryBuilder,
     GetRightHoldersQueryBuilder,
-} from "../../../clients/access-management-bff/connection/index.js";
+    ValidatePersonInputBuilder,
+} from "../../../../clients/access-management-bff/connection/index.js";
 import {
     ResourceClient,
     SearchResourcesQueryBuilder,
-} from "../../../clients/access-management-bff/resource/index.js";
+} from "../../../../clients/access-management-bff/resource/index.js";
 import {
     GetRolePermissionsQueryBuilder,
     RoleClient,
-} from "../../../clients/access-management-bff/role/index.js";
+} from "../../../../clients/access-management-bff/role/index.js";
 import {
     DelegateSingleRightsQueryBuilder,
     GetResourceDelegationsQueryBuilder,
@@ -30,70 +29,64 @@ import {
     GetSingleRightDelegationCheckQueryBuilder,
     RevokeSingleRightsQueryBuilder,
     SingleRightClient,
-} from "../../../clients/access-management-bff/single-right/index.js";
-import { UserClient } from "../../../clients/access-management-bff/user/index.js";
-import { PersonalTokenBuilder, PersonalTokenGenerator } from "../../../common-imports.js";
-import { getItemFromList, getNumberOfVUs, getOptions, parseCsvData, requireEnv, segmentData } from "../../../helpers.js";
-import { AltinnScopes, CreateScopeString } from "../../../scopes.js";
-import { GetAccessPackageDelegations } from "../../building-blocks/access-management-bff/access-package/index.js";
-import { SearchAccessPackages } from "../../building-blocks/access-management-bff/access-package/index.js";
+} from "../../../../clients/access-management-bff/single-right/index.js";
+import { UserClient } from "../../../../clients/access-management-bff/user/index.js";
+import { PersonalTokenBuilder, PersonalTokenGenerator } from "../../../../common-imports.js";
+import { getItemFromList, getNumberOfVUs, getOptions, parseCsvData, requireEnv, segmentData } from "../../../../helpers.js";
+import { AltinnScopes, CreateScopeString } from "../../../../scopes.js";
+import { GetAccessPackageDelegations } from "../../../building-blocks/access-management-bff/access-package/index.js";
+import { SearchAccessPackages } from "../../../building-blocks/access-management-bff/access-package/index.js";
+import { CreateRightHolder, DeleteReporteeConnection, GetRightHolders } from "../../../building-blocks/access-management-bff/connection/index.js";
+import { GetResourceOwners } from "../../../building-blocks/access-management-bff/resource/index.js";
+import { SearchResources } from "../../../building-blocks/access-management-bff/resource/index.js";
+import { GetRoles } from "../../../building-blocks/access-management-bff/role/index.js";
+import { GetRolePermissions } from "../../../building-blocks/access-management-bff/role/index.js";
+import { GetSingleRightDelegationCheck } from "../../../building-blocks/access-management-bff/single-right/index.js";
+import { GetResourceRights } from "../../../building-blocks/access-management-bff/single-right/index.js";
+import { GetResourceDelegations } from "../../../building-blocks/access-management-bff/single-right/index.js";
+import { GetRightsMeta } from "../../../building-blocks/access-management-bff/single-right/index.js";
+import { DelegateSingleRights, RevokeSingleRights } from "../../../building-blocks/access-management-bff/single-right/index.js";
 import {
-    GetOrgData,
-} from "../../building-blocks/access-management-bff/altinn-cdn/index.js";
-import { CreateRightHolder, DeleteReporteeConnection, GetRightHolders } from "../../building-blocks/access-management-bff/connection/index.js";
-import { GetResourceOwners, SearchResources } from "../../building-blocks/access-management-bff/resource/index.js";
-import { GetRolePermissions, } from "../../building-blocks/access-management-bff/role/index.js";
-import { GetRoles, } from "../../building-blocks/access-management-bff/role/index.js";
-import { RevokeSingleRights } from "../../building-blocks/access-management-bff/single-right/index.js";
-import { DelegateSingleRights } from "../../building-blocks/access-management-bff/single-right/index.js";
-import { GetRightsMeta } from "../../building-blocks/access-management-bff/single-right/index.js";
-import { GetResourceDelegations } from "../../building-blocks/access-management-bff/single-right/index.js";
-import { GetResourceRights } from "../../building-blocks/access-management-bff/single-right/index.js";
-import { GetSingleRightDelegationCheck } from "../../building-blocks/access-management-bff/single-right/index.js";
-import { GetIsHovedadmin } from "../../building-blocks/access-management-bff/user/index.js";
-import { getFromTo, getTokenOpts, } from "./commons.js";
-import { resourcesForOrg as resources } from "./custom-data.js";
+    GetIsHovedadmin,
+} from "../../../building-blocks/access-management-bff/user/index.js";
+import { getFromTo, getTokenOpts, } from "../commons.js";
+import { resourcesForUsers as resources } from "../custom-data.js";
 
 // Labels for different actions
-const getRightholdersLabel1a = { step: "1a. Get rightholders from org" };
-const postRightholderLabel = { step: "1b. Connecting organizations with PostRightholder" };
-const getRightholdersLabel1c = { step: "1c. Get rightholders org -> dagl" };
-const getRightholdersLabel1d = { step: "1d. Get rightholders from org" };
-const getRightholdersLabel1e = { step: "1e. Get rightholders from org to org" };
-const getIsHovedAdminLabel = { step: "1f. Get is hovedadmin for org" };
-const getRolePermissionsLabel = { step: "1g. Get role permissions for org to org" };
-const getDelegationsLabel = { step: "1h. Get delegations from org to org" };
-const getDelegatedResourcesLabel = { step: "1i. Get delegated resources from org to org" };
-const searchAccessPackagesLabel = { step: "1j. Search access packages for org" };
-const searchResourcesLabel = { step: "1k. Search resources for org" };
-const getResourceOwnersLabel = { step: "1l. Get resource owners for org" };
-const getOrganizationDataLabel = { step: "1m. Get organization data for org" };
+const getRightholdersLabel1a = { step: "1a. Get rightholders from user" };
+const postRightholderLabel = { step: "1b. Connecting users with PostRightholder" };
+const getRightholdersLabel1c = { step: "1c. Get rightholders from user to all" };
+const getRightholdersLabel1d = { step: "1d. Get rightholders from user to other user" };
+const getIsHovedAdminLabel = { step: "1f. Get is hovedadmin for user" };
+const getRolePermissionsLabel = { step: "1g. Get role permissions from user to user" };
+const getDelegationsLabel = { step: "1h. Get delegations from user to user" };
+const getDelegatedResourcesLabel = { step: "1i. Get delegated resources from user to user" };
+const searchAccessPackagesLabel = { step: "1j. Search access packages for user" };
+const searchResourcesLabel = { step: "1k. Search resources for person" };
+const getResourceOwnersLabel = { step: "1l. Get resource owners for user" };
 
-const searchAccessPackagesLabel2a = { step: "2a. Search access packages for org" };
+const searchAccessPackagesLabel2a = { step: "2a. Search resources for person" };
 const getRightsMetadataLabel2b = { step: "2b. Get rights metadata for resource" };
 const getDelegationCheckLabel = { step: "2c. Get delegation check for client delegation" };
-const postDelegationLabel = { step: "2d. Delegate serviceresource from org to org" };
-const getDelegatedResourcesLabel2d = { step: "2e. Get delegated resources for client delegation" };
-const getDelegationCheckLabel2e = { step: "2f. Get delegation check for client delegation after delegating" };
-const getDelegatedRightsForResourceLabel2f = { step: "2g. Get delegated rights for resource for client delegation" };
+const postDelegationLabel = { step: "2d. Delegate serviceresource from user to user" };
+const getDelegatedResourcesLabel2e = { step: "2e. Get delegated resources for user" };
+const getDelegationCheckLabel2f = { step: "2f. Get delegation check for user after delegating" };
+const getDelegatedRightsForResourceLabel2g = { step: "2g. Get delegated rights for resource for user" };
 
-const revokeSingleRightLabel = { step: "3a. Revoke single right for client delegation" };
-const getDelegatedResourcesLabel3b = { step: "3b. Get delegated resources for client delegation after revoking single right" };
-const getDelegationCheckLabel3c = { step: "3c. Get delegation check for client delegation after revoking single right" };
-const getRolePermissionsLabel3d = { step: "3d. Get role permissions for user to org" };
-const getRolePermissionsLabel3e = { step: "3e. Get role permissions for org to org" };
+const revokeSingleRightLabel = { step: "3a. Revoke single right for user" };
+const getDelegatedResourcesLabel3b = { step: "3b. Get delegated resources for user after revoking single right" };
+const getRolePermissionsLabel3d = { step: "3d. Get role permissions for user to user after revoking single right" };
 const getRoleMetaLabel3f = { step: "3f. Get role meta before revoking single right" };
-const getDelegationsLabel3g = { step: "3g. Get delegations from org to org" };
-const deleteRightholderConnectionLabel = { step: "3h. Delete rightholder connection between orgs" };
-const getRolePermissionsLabel3i = { step: "3i. Get role permissions for user to org after delete" };
-const getRolePermissionsLabel3j = { step: "3j. Get role permissions for org to org after delete" };
+const getDelegationsLabel3g = { step: "3g. Get delegations from user to user" };
+const deleteRightholderConnectionLabel = { step: "3h. Delete rightholder connection between users" };
+const getRolePermissionsLabel3j = { step: "3j. Get role permissions for user to user after delete" };
 const getRoleMetaLabel3k = { step: "3k. Get role meta after revoking single right" };
-const getDelegationsLabel3l = { step: "3l. Get delegations from org to org after delete" };
-const getRightholdersLabel3m = { step: "3m. Get rightholders from org after delete" };
+const getDelegationsLabel3l = { step: "3l. Get delegations from user to user after delete" };
+const getRightholdersLabel3m = { step: "3m. Get rightholders from user after delete" };
 
-const addUserGroup = { group: "1. Add organization as user to another organization" };
-const resourceDelegationGroup = { group: "2. Delegate a single resource to the added organization" };
-const cleanupGroup = { group: "3. Cleanup - delete delegation and connection" };
+const addUserGroup = { group: "1. ********* Connect ********" };
+const resourceDelegationGroup = { group: "2. ******** Delegate  ********" };
+const cleanupGroup = { group: "3. ********* Cleanup  *********" };
 
 const randomize = __ENV.RANDOMIZE ? __ENV.RANDOMIZE.toLowerCase() === "true" : true;
 
@@ -105,7 +98,6 @@ export const options = getOptions(
         postRightholderLabel,
         getRightholdersLabel1c,
         getRightholdersLabel1d,
-        getRightholdersLabel1e,
         getIsHovedAdminLabel,
         getRolePermissionsLabel,
         getDelegationsLabel,
@@ -113,27 +105,23 @@ export const options = getOptions(
         searchAccessPackagesLabel,
         searchResourcesLabel,
         getResourceOwnersLabel,
-        getOrganizationDataLabel,
 
         resourceDelegationGroup,
         searchAccessPackagesLabel2a,
         getRightsMetadataLabel2b,
         getDelegationCheckLabel,
         postDelegationLabel,
-        getDelegatedResourcesLabel2d,
-        getDelegationCheckLabel2e,
-        getDelegatedRightsForResourceLabel2f,
+        getDelegatedResourcesLabel2e,
+        getDelegationCheckLabel2f,
+        getDelegatedRightsForResourceLabel2g,
 
         cleanupGroup,
         revokeSingleRightLabel,
         getDelegatedResourcesLabel3b,
-        getDelegationCheckLabel3c,
         getRolePermissionsLabel3d,
-        getRolePermissionsLabel3e,
         getRoleMetaLabel3f,
         getDelegationsLabel3g,
         deleteRightholderConnectionLabel,
-        getRolePermissionsLabel3i,
         getRolePermissionsLabel3j,
         getRoleMetaLabel3k,
         getDelegationsLabel3l,
@@ -178,11 +166,6 @@ let roleApiClient = undefined;
 let resourceApiClient = undefined;
 
 /**
- * @type {AltinnCdnClient | undefined}
- */
-let altinnCdnApiClient = undefined;
-
-/**
  * Creates and caches the API clients used by the test.
  *
  * All clients share the same {@link PersonalTokenGenerator} instance.
@@ -195,7 +178,6 @@ let altinnCdnApiClient = undefined;
  * UserClient,
  * RoleClient,
  * ResourceClient,
- * AltinnCdnClient,
  * PersonalTokenGenerator
  * ]} The initialized API clients and token generator.
  */
@@ -209,6 +191,7 @@ function getClients() {
             .withTtl(3600)
             .withScopes(scopes)
             .build();
+
         tokenGenerator = new PersonalTokenGenerator(tokenOpts);
     }
 
@@ -236,10 +219,6 @@ function getClients() {
         resourceApiClient = new ResourceClient(__ENV.AM_UI_BASE_URL, tokenGenerator);
     }
 
-    if (altinnCdnApiClient == undefined) {
-        altinnCdnApiClient = new AltinnCdnClient(__ENV.AM_UI_BASE_URL, tokenGenerator);
-    }
-
     return [
         connectionsApiClient,
         accessPackageApiClient,
@@ -247,7 +226,6 @@ function getClients() {
         userApiClient,
         roleApiClient,
         resourceApiClient,
-        altinnCdnApiClient,
         tokenGenerator,
     ];
 }
@@ -255,12 +233,12 @@ function getClients() {
 /**
  * Setup function to segment data for VUs.
  *
- * @returns {object[][]} Organizations to delegate between, one slice per VU.
+ * @returns {object[][]} Users to delegate between, one slice per VU.
  */
 export function setup() {
     requireEnv(["ENVIRONMENT", "AM_UI_BASE_URL"]);
     const numberOfVUs = getNumberOfVUs();
-    const res = http.get(`https://raw.githubusercontent.com/Altinn/altinn-platform-validation-tests/refs/heads/main/K6/testdata/authentication/delegation/${__ENV.ENVIRONMENT}/single-service-org-org.csv`,
+    const res = http.get(`https://raw.githubusercontent.com/Altinn/altinn-platform-validation-tests/refs/heads/main/K6/testdata/authentication/delegation/${__ENV.ENVIRONMENT}/single-service-user-user.csv`,
         { tags: { action: "fetch-test-data" } });
     const segmentedData = segmentData(parseCsvData(res.body), numberOfVUs);
     return segmentedData;
@@ -269,7 +247,7 @@ export function setup() {
 /**
  * Main function executed by each VU.
  *
- * @param {object[][]} segmentedData Organizations to delegate between, one slice per VU.
+ * @param {object[][]} segmentedData Users to delegate between, one slice per VU.
  */
 export default function (segmentedData) {
     const [
@@ -279,11 +257,10 @@ export default function (segmentedData) {
         userApiClient,
         roleApiClient,
         resourceApiClient,
-        altinnCdnApiClient,
         tokenGenerator,
     ] = getClients();
 
-    // Get from and to organizations and resource for the current iteration
+    // Get from and to users and resource for the current iteration
     const { from, to } = getFromTo(segmentedData[exec.vu.idInTest - 1]);
     const resource = getItemFromList(resources, true);
 
@@ -291,24 +268,26 @@ export default function (segmentedData) {
     tokenGenerator.setTokenGeneratorOptions(getTokenOpts(from.userId, from.partyUuid));
 
     // Part 1.
-    // Add organization as user to another organization,
+    // Add user to auser,
     group(addUserGroup.group, function () {
         CreateRightHolder(
             connectionsApiClient,
-            from.orgUuid,
-            null,
-            new CreateRightHolderQueryBuilder()
-                .withRightholderPartyUuid(to.orgUuid)
+            from.partyUuid,
+            new ValidatePersonInputBuilder()
+                .withPersonIdentifier(to.ssn)
+                .withLastName(to.lastName)
                 .build(),
+            null,
             postRightholderLabel,
         );
         GetRightHolders(
             connectionsApiClient,
             new GetRightHoldersQueryBuilder()
-                .withParty(from.orgUuid)
-                .withFrom(from.orgUuid)
+                .withParty(from.partyUuid)
+                .withFrom(from.partyUuid)
+                .withTo(from.partyUuid)
                 .withIncludeClientDelegations(true)
-                .withIncludeAgentConnections(false)
+                .withIncludeAgentConnections(true)
                 .build(),
             getRightholdersLabel1a,
         );
@@ -316,8 +295,7 @@ export default function (segmentedData) {
             connectionsApiClient,
             new GetRightHoldersQueryBuilder()
                 .withParty(from.partyUuid)
-                .withFrom(from.orgUuid)
-                .withTo(from.partyUuid)
+                .withFrom(from.partyUuid)
                 .withIncludeClientDelegations(true)
                 .withIncludeAgentConnections(true)
                 .build(),
@@ -326,49 +304,39 @@ export default function (segmentedData) {
         GetRightHolders(
             connectionsApiClient,
             new GetRightHoldersQueryBuilder()
-                .withParty(from.orgUuid)
-                .withFrom(from.orgUuid)
-                .withIncludeClientDelegations(true)
-                .withIncludeAgentConnections(false)
-                .build(),
-            getRightholdersLabel1d,
-        );
-        GetRightHolders(
-            connectionsApiClient,
-            new GetRightHoldersQueryBuilder()
-                .withParty(from.orgUuid)
-                .withFrom(from.orgUuid)
-                .withTo(to.orgUuid)
+                .withParty(from.partyUuid)
+                .withFrom(from.partyUuid)
+                .withTo(to.partyUuid)
                 .withIncludeClientDelegations(true)
                 .withIncludeAgentConnections(true)
                 .build(),
-            getRightholdersLabel1e,
+            getRightholdersLabel1d,
         );
-        GetIsHovedadmin(userApiClient, from.orgUuid, getIsHovedAdminLabel);
+        GetIsHovedadmin(userApiClient, from.partyUuid, getIsHovedAdminLabel);
         GetRolePermissions(
             roleApiClient,
             new GetRolePermissionsQueryBuilder()
-                .withParty(from.orgUuid)
-                .withFrom(from.orgUuid)
-                .withTo(to.orgUuid)
+                .withParty(from.partyUuid)
+                .withFrom(from.partyUuid)
+                .withTo(to.partyUuid)
                 .build(),
             getRolePermissionsLabel,
         );
         GetAccessPackageDelegations(
             accessPackageApiClient,
             new GetAccessPackageDelegationsQueryBuilder()
-                .withParty(from.orgUuid)
-                .withFrom(from.orgUuid)
-                .withTo(to.orgUuid)
+                .withParty(from.partyUuid)
+                .withFrom(from.partyUuid)
+                .withTo(to.partyUuid)
                 .build(),
             getDelegationsLabel,
         );
         GetResourceDelegations(
             singleRightsApiClient,
             new GetResourceDelegationsQueryBuilder()
-                .withParty(from.orgUuid)
-                .withFrom(from.orgUuid)
-                .withTo(to.orgUuid)
+                .withParty(from.partyUuid)
+                .withFrom(from.partyUuid)
+                .withTo(to.partyUuid)
                 .build(),
             getDelegatedResourcesLabel,
         );
@@ -376,7 +344,7 @@ export default function (segmentedData) {
             accessPackageApiClient,
             new SearchAccessPackagesQueryBuilder()
                 .withSearchString("")
-                .withTypeName("organisasjon")
+                .withTypeName("person")
                 .build(),
             searchAccessPackagesLabel,
         );
@@ -391,17 +359,16 @@ export default function (segmentedData) {
             searchResourcesLabel,
         );
         GetResourceOwners(resourceApiClient, null, getResourceOwnersLabel);
-        GetOrgData(altinnCdnApiClient, getOrganizationDataLabel);
     });
 
     // Part 2.
-    // Delegate a single resource to the added organization and verify delegation
+    // Delegate a single resource to the added user and verify delegation
     group(resourceDelegationGroup.group, function () {
         SearchAccessPackages(
             accessPackageApiClient,
             new SearchAccessPackagesQueryBuilder()
                 .withSearchString(resource.searchTerm)
-                .withTypeName("organisasjon")
+                .withTypeName("person")
                 .build(),
             searchAccessPackagesLabel2a,
         );
@@ -415,7 +382,7 @@ export default function (segmentedData) {
         GetSingleRightDelegationCheck(
             singleRightsApiClient,
             new GetSingleRightDelegationCheckQueryBuilder()
-                .withFrom(from.orgUuid)
+                .withFrom(from.partyUuid)
                 .withResource(resource.resourceId)
                 .build(),
             getDelegationCheckLabel,
@@ -423,9 +390,9 @@ export default function (segmentedData) {
         DelegateSingleRights(
             singleRightsApiClient,
             new DelegateSingleRightsQueryBuilder()
-                .withParty(from.orgUuid)
-                .withFrom(from.orgUuid)
-                .withTo(to.orgUuid)
+                .withParty(from.partyUuid)
+                .withFrom(from.partyUuid)
+                .withTo(to.partyUuid)
                 .withResourceId(resource.resourceId)
                 .build(),
             getRights(rightsMeta),
@@ -434,42 +401,42 @@ export default function (segmentedData) {
         GetResourceDelegations(
             singleRightsApiClient,
             new GetResourceDelegationsQueryBuilder()
-                .withParty(from.orgUuid)
-                .withFrom(from.orgUuid)
-                .withTo(to.orgUuid)
+                .withParty(from.partyUuid)
+                .withFrom(from.partyUuid)
+                .withTo(to.partyUuid)
                 .build(),
-            getDelegatedResourcesLabel2d,
+            getDelegatedResourcesLabel2e,
         );
         GetSingleRightDelegationCheck(
             singleRightsApiClient,
             new GetSingleRightDelegationCheckQueryBuilder()
-                .withFrom(from.orgUuid)
+                .withFrom(from.partyUuid)
                 .withResource(resource.resourceId)
                 .build(),
-            getDelegationCheckLabel2e,
+            getDelegationCheckLabel2f,
         );
         GetResourceRights(
             singleRightsApiClient,
             new GetResourceRightsQueryBuilder()
-                .withParty(from.orgUuid)
-                .withFrom(from.orgUuid)
-                .withTo(to.orgUuid)
+                .withParty(from.partyUuid)
+                .withFrom(from.partyUuid)
+                .withTo(to.partyUuid)
                 .withResourceId(resource.resourceId)
                 .build(),
-            getDelegatedRightsForResourceLabel2f,
+            getDelegatedRightsForResourceLabel2g,
         );
     });
 
     // Part 3.
     // Revoke the delegated resource and verify that the delegation has been removed,
-    // then clean up by deleting the rightholder connection between the organizations and verify deletion
+    // then clean up by deleting the rightholder connection between the users and verify deletion
     group(cleanupGroup.group, function () {
         RevokeSingleRights(
             singleRightsApiClient,
             new RevokeSingleRightsQueryBuilder()
-                .withParty(from.orgUuid)
-                .withFrom(from.orgUuid)
-                .withTo(to.orgUuid)
+                .withParty(from.partyUuid)
+                .withFrom(from.partyUuid)
+                .withTo(to.partyUuid)
                 .withResourceId(resource.resourceId)
                 .build(),
             revokeSingleRightLabel,
@@ -477,64 +444,37 @@ export default function (segmentedData) {
         GetResourceDelegations(
             singleRightsApiClient,
             new GetResourceDelegationsQueryBuilder()
-                .withParty(from.orgUuid)
-                .withFrom(from.orgUuid)
-                .withTo(to.orgUuid)
+                .withParty(from.partyUuid)
+                .withFrom(from.partyUuid)
+                .withTo(to.partyUuid)
                 .build(),
             getDelegatedResourcesLabel3b,
-        );
-        GetSingleRightDelegationCheck(
-            singleRightsApiClient,
-            new GetSingleRightDelegationCheckQueryBuilder()
-                .withFrom(from.orgUuid)
-                .withResource(resource.resourceId)
-                .build(),
-            getDelegationCheckLabel3c,
-        );
-        GetResourceRights(
-            singleRightsApiClient,
-            new GetResourceRightsQueryBuilder()
-                .withParty(from.orgUuid)
-                .withFrom(from.orgUuid)
-                .withTo(to.orgUuid)
-                .withResourceId(resource.resourceId)
-                .build(),
-            getRolePermissionsLabel,
         );
         GetRolePermissions(
             roleApiClient,
             new GetRolePermissionsQueryBuilder()
                 .withParty(from.partyUuid)
-                .withFrom(from.orgUuid)
-                .withTo(from.partyUuid)
+                .withFrom(from.partyUuid)
+                .withTo(to.partyUuid)
                 .build(),
             getRolePermissionsLabel3d,
-        );
-        GetRolePermissions(
-            roleApiClient,
-            new GetRolePermissionsQueryBuilder()
-                .withParty(from.orgUuid)
-                .withFrom(from.orgUuid)
-                .withTo(to.orgUuid)
-                .build(),
-            getRolePermissionsLabel3e,
         );
         GetRoles(roleApiClient, getRoleMetaLabel3f);
         GetAccessPackageDelegations(
             accessPackageApiClient,
             new GetAccessPackageDelegationsQueryBuilder()
-                .withParty(from.orgUuid)
-                .withFrom(from.orgUuid)
-                .withTo(to.orgUuid)
+                .withParty(from.partyUuid)
+                .withFrom(from.partyUuid)
+                .withTo(to.partyUuid)
                 .build(),
             getDelegationsLabel3g,
         );
         DeleteReporteeConnection(
             connectionsApiClient,
             new DeleteReporteeConnectionQueryBuilder()
-                .withParty(from.orgUuid)
-                .withFrom(from.orgUuid)
-                .withTo(to.orgUuid)
+                .withParty(from.partyUuid)
+                .withFrom(from.partyUuid)
+                .withTo(to.partyUuid)
                 .build(),
             deleteRightholderConnectionLabel,
         );
@@ -542,17 +482,8 @@ export default function (segmentedData) {
             roleApiClient,
             new GetRolePermissionsQueryBuilder()
                 .withParty(from.partyUuid)
-                .withFrom(from.orgUuid)
-                .withTo(from.partyUuid)
-                .build(),
-            getRolePermissionsLabel3i,
-        );
-        GetRolePermissions(
-            roleApiClient,
-            new GetRolePermissionsQueryBuilder()
-                .withParty(from.orgUuid)
-                .withFrom(from.orgUuid)
-                .withTo(to.orgUuid)
+                .withFrom(from.partyUuid)
+                .withTo(to.partyUuid)
                 .build(),
             getRolePermissionsLabel3j,
         );
@@ -560,24 +491,23 @@ export default function (segmentedData) {
         GetAccessPackageDelegations(
             accessPackageApiClient,
             new GetAccessPackageDelegationsQueryBuilder()
-                .withParty(from.orgUuid)
-                .withFrom(from.orgUuid)
-                .withTo(to.orgUuid)
+                .withParty(from.partyUuid)
+                .withFrom(from.partyUuid)
+                .withTo(to.partyUuid)
                 .build(),
             getDelegationsLabel3l,
         );
         GetRightHolders(
             connectionsApiClient,
             new GetRightHoldersQueryBuilder()
-                .withParty(from.orgUuid)
-                .withFrom(from.orgUuid)
+                .withParty(from.partyUuid)
+                .withFrom(from.partyUuid)
                 .withIncludeClientDelegations(true)
-                .withIncludeAgentConnections(false)
+                .withIncludeAgentConnections(true)
                 .build(),
             getRightholdersLabel3m,
         );
     });
-
 }
 
 /**
