--- K6/api/tests/access-management-bff/smoke.yaml
@@ -1,83 +1,0 @@
-namespace: access-management-bff
-base_dir: K6/api/tests/access-management-bff
-test_definitions:
-  - test_scope: access-management-bff
-    test_file: client-admin.js
-    contexts:
-      - environment: yt01
-        test_type:
-          type: smoke
-          enabled: true
-          config_file: conf.json
-      - environment: at23
-
-  - test_scope: access-management-bff
-    test_file: fullmakt-user-user.js
-    contexts:
-      - environment: yt01
-        test_type:
-          type: smoke
-          enabled: true
-          config_file: conf.json
-      - environment: at23
-
-  - test_scope: access-management-bff
-    test_file: fullmakt-org-org.js
-    contexts:
-      - environment: yt01
-        test_type:
-          type: smoke
-          enabled: true
-          config_file: conf.json
-      - environment: at23
-
-  - test_scope: access-management-bff
-    test_file: single-service-org-org.js
-    contexts:
-      - environment: yt01
-        test_type:
-          type: smoke
-          enabled: true
-          config_file: conf.json
-      - environment: at23
-
-  - test_scope: access-management-bff
-    test_file: single-service-user-user.js
-    contexts:
-      - environment: yt01
-        test_type:
-          type: smoke
-          enabled: true
-          config_file: conf.json
-      - environment: at23
-
-  - test_scope: access-management-bff
-    test_file: instance-delegation-user-user.js
-    contexts:
-      - environment: at23
-        test_type:
-          type: smoke
-          enabled: true
-          config_file: conf.json
-
-  - test_scope: access-management-bff
-    test_file: instance-delegation-org-user.js
-    contexts:
-      - environment: at23
-        test_type:
-          type: smoke
-          enabled: true
-          config_file: conf.json
-
-  - test_scope: access-management-bff
-    test_file: delegation-export.js
-    contexts:
-      - environment: yt01
-        test_type:
-          type: smoke
-          enabled: true
-          config_file: conf.json
-      - environment: at23
-        test_type:
-          type: smoke
-          enabled: true
